+++ monitor.py
@@ -859,23 +859,14 @@ class WorkReaper(ReaperMonitor):
     """
     MODEL_CLASS = Work
 
-    def __init__(self, *args, **kwargs):
-        from external_search import ExternalSearchIndex
-        search_index_client = kwargs.pop('search_index_client', None)
-        super(WorkReaper, self).__init__(*args, **kwargs)
-        self.search_index_client = (
-            search_index_client or ExternalSearchIndex(self._db)
-        )
-
     def query(self):
         return self._db.query(Work).outerjoin(Work.license_pools).filter(
             LicensePool.id==None
         )
 
     def delete(self, work):
-        """Delete work from elasicsearch and database."""
-        self.search_index_client.remove_work(work)
-        self._db.delete(work)
+        """Delete work from elasticsearch and database."""
+        work.delete()
 
 ReaperMonitor.REGISTRY.append(WorkReaper)
